travail sur la customisation des routes des voyages en cours

+++ src/main/java/com/annulation_reservation_voyage/annulation_reservation_voyage/models/Historique.java
@@ -25,6 +25,6 @@ public class Historique {
   private LocalDateTime dateAnnulation;
   private String causeAnnulation;
   private String origineAnnulation;
-  private float tauxAnnulation;
-  private float compensation;
+  private double tauxAnnulation;
+  private double compensation;
 }
