+++ src/main/java/com/annulation_reservation_voyage/annulation_reservation_voyage/models/PolitiqueAnnulation.java
@@ -19,5 +19,5 @@ public class PolitiqueAnnulation {
   private UUID idPolitique;
   private LocalDateTime dateDebut;
   private LocalDateTime dateFin;
-  private float taux;
+  private double taux;
 }
